Properly cast buffer and latency times when setting properties
This fix a crash on 32 bits system as gcc can't figure out the right size of
the value.

Thanks to Raluca-Elena Podiuc for her help in debugging this.

https://bugzilla.gnome.org/show_bug.cgi?id=665171

+++ src/empathy-audio-sink.c
@@ -222,7 +222,10 @@ create_sink (EmpathyGstAudioSink *self)
 
   /* Set latency (buffering on the PulseAudio side) of 40ms and transfer data
    * in 10ms chunks */
-  g_object_set (sink, "buffer-time", 40000, "latency-time", 10000, NULL);
+  g_object_set (sink,
+      "buffer-time", (gint64) 40000,
+      "latency-time", (gint64) 10000,
+      NULL);
 
   return sink;
 }
